requirements.txt
# Database
mysqlclient==2.2.4

# Security & Auth
cryptography==43.0.0
PyJWT==2.8.0
bcrypt==4.1.2
PyNaCl==1.5.0

# HTTP Requests
requests==2.32.3
httpx==0.27.0

# File Upload & Env Management
python-dotenv==1.0.1
python-multipart==0.0.9

# Jinja Templating (if needed)
Jinja2==3.1.4
